+++ tests/testthat/test-exec_paths.R
@@ -5,7 +5,8 @@ exec_path_funs <- list(
   pip = pip_exec,
   chrome = chrome_exec,
   firefox = firefox_exec,
-  libreoffice = libreoffice_exec
+  libreoffice = libreoffice_exec,
+  air = air_exec
 )
 
 test_that("*_exec() returns the path found by exec_locate", {