+++ NEWS.md
@@ -1,5 +1,6 @@
 # locatexec 0.1.2
 
+* Add detection of the 'air' executable (R formatter).
 * Fix `min_version_available(version=...)`
 * Set `NODE_DISABLE_COMPILE_CACHE=1` when calling node/npm to avoid leaving a `node-compile-cache` directory in the temp directory.